move the MainImage component to the components folder
