+++ src/components/MainImage.js
