+++ src/components/MainImage.sc.js
