+++ src/components/redirect_notification/SupportedNotification.js
@@ -4,7 +4,7 @@ import Modal from "../modal_shared/Modal";
 import Header from "../modal_shared/Header";
 import Heading from "../modal_shared/Heading";
 import Main from "../modal_shared/Main";
-import MainImage from "../modal_shared/MainImage";
+import MainImage from "../MainImage";
 import Footer from "../modal_shared/Footer";
 import ButtonContainer from "../modal_shared/ButtonContainer";
 import Checkbox from "../modal_shared/Checkbox";
