+++ src/pages/ExtensionInstalled.js
@@ -4,7 +4,7 @@ import InfoPage from "./info_page_shared/InfoPage";
 import Header from "./info_page_shared/Header";
 import Heading from "./info_page_shared/Heading";
 import Main from "./info_page_shared/Main";
-import MainImage from "./info_page_shared/MainImage";
+import MainImage from "../components/MainImage";
 import ButtonContainer from "./info_page_shared/ButtonContainer";
 import Footer from "./info_page_shared/Footer";
 import Button from "./info_page_shared/Button";
