+++ src/pages/InstallExtension.js
@@ -6,7 +6,7 @@ import ButtonContainer from "./info_page_shared/ButtonContainer";
 import Footer from "./info_page_shared/Footer";
 import Button from "./info_page_shared/Button";
 import { getExtensionInstallationLinks } from "../utils/misc/extensionInstallationLinks";
-import MainImage from "./info_page_shared/MainImage";
+import MainImage from "../components/MainImage";
 import FileDownloadOutlinedIcon from "@mui/icons-material/FileDownloadOutlined";
 
 export default function InstallExtension() {
