--- src/pages/info_page_shared/MainImage.js
@@ -1,5 +1,0 @@
-import * as s from "./MainImage.sc";
-
-export default function MainImage({ src }) {
-  return <s.MainImage src={src} alt="" />;
-}
