--- src/pages/info_page_shared/MainImage.sc.js
@@ -1,9 +1,0 @@
-import styled from "styled-components";
-
-const MainImage = styled.img`
-  height: 100%;
-  width: 100%;
-  object-fit: contain;
-`;
-
-export { MainImage };
